+++ front/app/containers/App/ModalQueue/modals/UserSessionRecording/ModalManager.tsx
